Added Issue Latency.
Instructions should now correctly track their issue latency (referred to
as reciprocal for "reciprocal throughput").  This means that an LU will
keep its 'current' pointer until the instruction has executed for N cycles,
where N is the reciprocal throughput of the instruction.

This fucntionality was added to the instruction_t class instead of the
lu_t class because the issue latency is different for different
instructions executing on a single LU.